feat: add product validations

+++ src/routes/product-routes.js
@@ -1,5 +1,6 @@
 import { ProductRepository } from "../repository/index.js";
 import { Router } from "express";
+import { uuidSchema, productSchema } from "../schemas/index.js";
 
 export const productRouter = Router();
 
@@ -12,13 +13,25 @@ productRouter.get("/", async (req, res) => {
 productRouter.get("/:id", async (req, res) => {
   const { id } = req.params;
 
+  const validation = uuidSchema.safeParse(id);
+
+  if (!validation.success) {
+    return res.status(400).send(validation.error.errors);
+  }
+
   const result = await new ProductRepository().getProductById(id);
+
+  if (!result.length) return res.status(404).send("Produto não encontrado.");
 
   return res.status(200).send(result);
 });
 
 productRouter.post("/", async (req, res) => {
   const { body } = req;
+
+  const validation = productSchema.safeParse(body);
+
+  if (!validation.success) return res.status(400).send(validation.error.errors);
 
   const productColumns = ["name", "price_in_cents", "size"];
 
@@ -37,6 +50,21 @@ productRouter.put("/:id", async (req, res) => {
   const { id } = req.params;
   const { body } = req;
 
+  const idValidation = uuidSchema.safeParse(id);
+
+  if (!idValidation.success)
+    return res.status(400).send(idValidation.error.errors);
+
+  const searchedProduct = await new ProductRepository().getProductById(id);
+
+  if (!searchedProduct.length)
+    return res.status(404).send("Produto não encontrado.");
+
+  const bodyValidation = productSchema.safeParse(body);
+
+  if (!bodyValidation.success)
+    return res.status(400).send(bodyValidation.error.errors);
+
   const columns = ["name", "price_in_cents", "size"];
 
   const values = columns.map((column) => body[column]);
@@ -48,6 +76,15 @@ productRouter.put("/:id", async (req, res) => {
 
 productRouter.delete("/:id", async (req, res) => {
   const { id } = req.params;
+
+  const validation = uuidSchema.safeParse(id);
+
+  if (!validation.success) return res.status(400).send(validation.error.errors);
+
+  const searchedProduct = await new ProductRepository().getProductById(id);
+
+  if (!searchedProduct.length)
+    return res.status(404).send("Produto não encontrado.");
 
   await new ProductRepository().deleteProductById(id);
